Partially implement URL rewrites for all file documents.
This URL rewrite currently only works for public accessible docuemnts.
(Enabling it for further access groups will follow.) By this commit
a new upload folder is created and for each access to a file a new
symlink from that folder to the actual storage folder is generated.
The storage folder then can be made non-readable (by htaccess).